requirements.txt
fastapi
uvicorn
streamlit
sqlalchemy
pydantic
requests
python-dotenv
groq
psycopg2-binary
